+++ lib/generators/cms/component/editing/editing_generator.rb
@@ -26,6 +26,33 @@ module Cms
           directory('app')
         end
 
+        def update_application_js
+          file = 'app/assets/javascripts/application.js'
+          insert_point = "//= require infopark_rails_connector"
+
+          data = []
+
+          data << ''
+          data << '//= require editing'
+
+          data = data.join("\n")
+
+          insert_into_file(file, data, after: insert_point)
+        end
+
+        def update_application_css
+          file = 'app/assets/stylesheets/application.css'
+          insert_point = '*= require infopark_rails_connector'
+
+          data = []
+          data << ''
+          data << ' *= require editing'
+
+          data = data.join("\n")
+
+          insert_into_file(file, data, after: insert_point)
+        end
+
         def run_generator_for_selected_editor
           Rails::Generators.invoke("cms:component:editing:#{editor}", [], behavior: behavior)
         end